Move `getLocationGroupedOptions` to server file
