+++ packages/trpc/server/routers/loggedInViewer/locationOptions.handler.ts
@@ -1,4 +1,4 @@
-import { getLocationGroupedOptions } from "@calcom/app-store/utils";
+import { getLocationGroupedOptions } from "@calcom/app-store/server";
 import { getTranslation } from "@calcom/lib/server/i18n";
 import type { TrpcSessionUser } from "@calcom/trpc/server/trpc";
